requirements.txt
# Web scraping
requests>=2.28.1
beautifulsoup4>=4.11.1
lxml>=4.9.1

# Database
SQLAlchemy>=1.4.41

# Reddit API
praw>=7.6.1
prawcore>=2.3.0

# Natural Language Processing
nltk>=3.7
vaderSentiment>=3.3.2

# Web application
Flask>=2.2.2
Jinja2>=3.1.2
Werkzeug>=2.2.2

# Utility packages
python-dotenv>=0.21.0
schedule>=1.1.0

# Date handling
pytz>=2022.4

# For development/testing
pytest>=7.1.3